add new files for testing posts and authors
