+++ backend/tests/test_posts.py
@@ -18,10 +18,7 @@ class TestPostViewSet(APITestCase):
         #     'john', 'john@snow.com', 'johnpassword')
 
         # Why can't I use APIClient() out here?
-        self.client = APIClient(
-            HTTP_HOST='127.0.0.1:8000',
-            ALLOWED_HOSTS='127.0.0.1'
-        )
+        self.client = APIClient()
 
         # self.data = {'displayName': 'John',
         #              "github": "https://github.com/johnSnow",
@@ -29,18 +26,18 @@ class TestPostViewSet(APITestCase):
         #              'username': 'John'
         #              }
 
-        self.author = Author.objects.create(**get_test_post_fields())
-        self.post_test = Post.objects.create(
-            **get_test_post_fields(), author=self.author)
+    #     self.author = Author.objects.create(**get_test_post_fields())
+    #     self.post_test = Post.objects.create(
+    #         **get_test_post_fields(), author=self.author)
 
-        self.create_post_url = reverse('posts_create', kwargs=self.author.id)
+    #     self.create_post_url = reverse('posts_create', kwargs=self.author.id)
 
-    def test_create_post(self):
-        """Testing for creation of a post made by an author
-        """
-        response = self.client.post(self.create_post_url)
+    # def test_create_post(self):
+    #     """Testing for creation of a post made by an author
+    #     """
+    #     response = self.client.post(self.create_post_url)
 
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+    #     self.assertEqual(response.status_code, status.HTTP_201_CREATED)
 
     # def test_create_user_payload_check(self):
     #     response = self.client.post(REGISTER_USER_URL, self.data)
